multiple-masters + disabled seed_master
Commented out `seed_master` just to manually accept the keys again.
For the second master, the minions throw

> Error while bringing up minion for multi-master. Is master at
192.168.50.11 responding?

errors which led me to examine whether it could have had something to do
with not explicitly accepting the keys for the master2 vm. It doesn't
seem to matter in retrospect. Anyways. Changing the minions to use the
IP address of the 2nd master (192.168.50.11) will fail. Even though the
configuration for the second master is identical to the first. Not sure
if something fails due to some crytographic operations that perhaps use
hashes that depend on machine-specific inputs. I would guess that is a
bullshit idea from me anyways since I've destroyed the entire setup
numerous times and respawned it, resulting to different VM's and every
time the master1 seems to function, but master2 doesn't.

At this moment in time, all minions just connect to master1. Change the
minion configuration for perhaps just one minion to use the IP of
master2 and observe how it fails.